Make the scripts work properly on a clean install

+++ applyPatches.py
@@ -2,6 +2,7 @@
 import os
 from os import path
 from scripts import core
+from scripts.core import eprint
 from shutil import copytree, rmtree
 
 core.setup()
@@ -17,8 +18,9 @@ if not path.exists("minecraft/src/net/minecraft"):
     eprint("Decompiled sources not found")
     exit(1)
 
-print("Removing existing sources")
-rmtree("src/main/java/net/minecraft")
+if path.exists("src/main/java/net/minecraft"):
+    print("Removing existing sources")
+    rmtree("src/main/java/net/minecraft")
 
 print("Copying decompiled sources")
 copytree("minecraft/src/net/minecraft", "src/main/java/net/minecraft")
